requirements.txt
asgiref==3.4.1
attrs==21.2.0
colorama==0.4.4
Django==3.2.7
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.1
redgreenunittest==0.1.1
sqlparse==0.4.1
toml==0.10.2
